refactor(oxfmt): Manage TOML options under `oxfmtrc.rs` as well (#18021)
Now, all options are managed under `oxfmtrc`.

Additionally, it improves performance a bit when handling multiple TOML files.
